Search input strategy
Provide a search input field alongside the task type filters, perform
searches with hastemplate appended. The idea is to mimic the search UI
from MobileFrontend and provide dynamic feedback to the user until
they find tasks they are interested in.

We could consider using topics to kick off some of the searches,
e.g. selecting "Filosofie" from the MenuTagMultiselectWidget would
prefill the serach input with "Filosofie" and execute a search.